分類帳完成大部份，可將初始餘額算入，未加入月份累積金額數；完成財務報告HelloWorld

+++ app/src/main/java/com/vincent/acnt/data/Entry.java
@@ -13,7 +13,7 @@ public class Entry implements Serializable {
 
     public Entry() {}
 
-    public Entry(int date, String memo, String ps) {
+    public Entry(long date, String memo, String ps) {
         this.subjects = new ArrayList<>();
         this.date = date;
         this.memo = memo;
